+++ CHANGELOG.md
@@ -2,6 +2,17 @@
 
 All notable changes to this project will be documented in this file. This project adheres to [Semantic Versioning](http://semver.org/).
 
+## [0.1.33] - 2025-10-22
+
+## Added
+- Task cluster server roles.
+- Sharing settings.
+- Object storage quotas for email and groupware.
+
+### Changed
+
+### Fixed
+
 ## [0.1.32] - 2025-07-28
 
 ## Added